Move to App\Model\Parameters namespace
