+++ src/Model/Parameters/JsonParameters.php
@@ -1,7 +1,7 @@
 <?php
 declare(strict_types=1);
 
-namespace App\Model\ProbeArgument;
+namespace App\Model\Parameters;
 
 abstract class JsonParameters implements JsonParametersInterface, \ArrayAccess
 {
