+++ src/Model/Parameters/NullParameters.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace App\Model\ProbeArgument;
+namespace App\Model\Parameters;
 
 class NullParameters extends JsonParameters
 {
