+++ src/Model/Parameters/PingParameters.php
@@ -1,7 +1,7 @@
 <?php
 declare(strict_types=1);
 
-namespace App\Model\ProbeArgument;
+namespace App\Model\Parameters;
 
 class PingParameters extends JsonParameters
 {
