+++ tests/App/Model/ProbeArgument/HttpParametersTest.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace App\Tests\App\Model\ProbeArgument;
 
-use App\Model\ProbeArgument\HttpParameters;
+use App\Model\Parameters\HttpParameters;
 use PHPUnit\Framework\TestCase;
 
 class HttpParametersTest extends TestCase
